Lesson 3 of 4

+++ conFusion/www/js/controllers.js
@@ -221,20 +221,34 @@ angular.module('conFusion.controllers', [])
 
   }])
 
-  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+  .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate',
+    '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL,
+                                                          $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
-    $scope.baseURL = baseURL;
-    $scope.shouldShowDelete = false;
 
-    $scope.favorites = favoriteFactory.getFavorites();
+      $scope.baseURL = baseURL;
+      $scope.shouldShowDelete = false;
 
-    $scope.dishes = menuFactory.getDishes().query(
-      function (response) {
-        $scope.dishes = response;
-      },
-      function (response) {
-        $scope.message = "Error: " + response.status + " " + response.statusText;
+      $ionicLoading.show({
+        template: '<ion-spinner></ion-spinner> Loading...'
       });
+
+      $scope.favorites = favoriteFactory.getFavorites();
+
+      $scope.dishes = menuFactory.getDishes().query(
+        function (response) {
+          $scope.dishes = response;
+          $timeout(function () {
+            $ionicLoading.hide();
+          }, 1000);
+        },
+        function (response) {
+          $scope.message = "Error: " + response.status + " " + response.statusText;
+          $timeout(function () {
+            $ionicLoading.hide();
+          }, 1000);
+        });
+
     console.log($scope.dishes, $scope.favorites);
 
     $scope.toggleDelete = function () {
@@ -242,12 +256,28 @@ angular.module('conFusion.controllers', [])
       console.log($scope.shouldShowDelete);
     }
 
-    $scope.deleteFavorite = function (index) {
+      $scope.deleteFavorite = function (index) {
 
-      favoriteFactory.deleteFromFavorites(index);
-      $scope.shouldShowDelete = false;
+        var confirmPopup = $ionicPopup.confirm({
+          title: 'Confirm Delete',
+          template: 'Are you sure you want to delete this item?'
+        });
 
-    }}])
+        confirmPopup.then(function (res) {
+          if (res) {
+            console.log('Ok to delete');
+            favoriteFactory.deleteFromFavorites(index);
+          } else {
+            console.log('Canceled delete');
+          }
+        });
+
+        $scope.shouldShowDelete = false;
+
+      }
+
+    }])
+
   .filter('favoriteFilter', function () {
     return function (dishes, favorites) {
       var out = [];
